==> notebooks/4.2_Viz_GEX_Data.ipynb <==
# 4.2 Visualize GEX Data

import pandas as pd
import numpy as np
from copy import deepcopy
from ast import literal_eval as make_tuple

import matplotlib.pyplot as plt
%matplotlib inline 

from clustergrammer2 import net
import helper_functions as hf

df = pd.read_parquet('../data/10k_pbmc_feature_v3-chem_v3.0.0-cr/processed_data/gex-cat_250-var.parquet')
df.columns = [make_tuple(x) for x in df.columns]

df.shape

net.load_df(df)
net.normalize(axis='row', norm_type='zscore')
net.widget()
