==> notebooks/5.0_Pre-Process_2-Million_Mouse_Atlas.ipynb <==
# 5.0 Pre-Process 2-Million Mouse Atlas
https://oncoscape.v3.sttrcancer.org/atlas.gs.washington.edu.mouse.rna/downloads

import pandas as pd
from glob import glob
import os

all_samples = sorted(glob('../data/big_data/cao_2million-cell_2019_61-embryo_parquet_files/*'))
len(all_samples)

def make_dir(directory):
    import os
    if not os.path.exists(directory):
        os.mkdir(directory)

### Calc Variance of All Genes Across Samples

def calc_gene_mean(inst_sample):
    df_gex = pd.read_parquet(inst_sample + '/gex.parquet')
    print(df_gex.shape)
    inst_mean = df_gex.mean(axis=1)
    return inst_mean

df_meta.columns.tolist()

### Drop cells from doublet derived sub-clusters

df_meta_ini = pd.read_parquet(all_samples[0] + '/meta_cell.parquet')
ser_doublet = df_meta_ini['doublet_cluster']
ser_doublet = ser_doublet[ser_doublet == False]
keep_cells = ser_doublet.index.tolist()
df_meta = df_meta_ini.loc[keep_cells]
df_meta.shape

### Find the top variable genes in a representative embryo

df_gex = pd.read_parquet(all_samples[0] + '/gex.parquet', columns=keep_cells)
print(df_gex.shape)
inst_mean = df_gex.mean(axis=1)
top_genes = inst_mean.sort_values(ascending=False).index.tolist()[:10000]
keep_genes = df_gex.loc[top_genes].var(axis=1).sort_values(ascending=False).index.tolist()[:5000]

### Make Top 5K Gene Variance Versions of Embryo DataSets
The top 5,000 variable genes were defined based on embryo-1E9.5 gene expression data

base_dir = '../data/cao_2million-cell_2019_61-embryo_parquet_files_binder/'

for inst_sample_path in all_samples:
    inst_sample = inst_sample_path.split('/')[-1]
    new_sample_dir = base_dir + inst_sample

    # drop doublets
    df_meta_ini = pd.read_parquet(inst_sample_path + '/meta_cell.parquet')
    ser_doublet = df_meta_ini['doublet_cluster']
    ser_doublet = ser_doublet[ser_doublet == False]
    keep_cells = ser_doublet.index.tolist()
    df_meta = df_meta_ini.loc[keep_cells]    

    # load gene expression 
    df_gex = pd.read_parquet(inst_sample_path + '/gex.parquet', columns=keep_cells).loc[keep_genes]
    
    # save filtered data for mybinder
    make_dir(new_sample_dir)
    df_meta.to_parquet(new_sample_dir + '/meta_cell.parquet')    
    df_gex.to_parquet(new_sample_dir + '/gex.parquet')
    
    # check file size (MB)
    statinfo = os.stat(new_sample_dir + '/gex.parquet')
    print(inst_sample, df_gex.shape, statinfo.st_size/(1000000))

### Verifying ~1.3 Million remaining Cells
Compare to processed data after removing cells from doublet derived clusters (https://oncoscape.v3.sttrcancer.org/atlas.gs.washington.edu.mouse.rna/downloads).

total_cells = 0
new_samples = sorted(glob('../data/big_data/cao_2million-cell_2019_61-embryo_parquet_files_binder/*'))
for inst_sample in new_samples:

    df_meta = pd.read_parquet(inst_sample + '/meta_cell.parquet')
    total_cells = total_cells + df_meta.shape[0]
print(total_cells)
