==> notebooks/5.2_Viz_2M-Mouse-Atlas_Downsampled.ipynb <==
# 5.2 Viz 2M-Mouse-Atlas Downsampled

from clustergrammer2 import net
df = {}

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
from ast import literal_eval as make_tuple

df_ds = pd.read_parquet('../data/cao_2million-cell_2019_61-embryo_downsample/cao_embryo_cell-type_downsample.parquet')
df_ds.columns = [make_tuple(x) for x in df_ds.columns.tolist()]
df_ds.shape

ser_cell_types = pd.Series([x[1].split(': ')[1] for x in df_ds.columns.tolist()])

### Cell Type Distributions

cell_type_dist = pd.read_csv('../data/cao_2million-cell_2019_61-embryo_downsample/cell_type_dist.txt', sep='\t', index_col=0, header=None)
print(cell_type_dist)

#### Total number of cells in downsampled view

cell_type_dist.sum().get_values()[0]

# Visualize Downsampled Data

net.load_df(df_ds)
net.filter_N_top(inst_rc='row', N_top=250, rank_type='var')
net.normalize(axis='row', norm_type='zscore')
net.clip(-5,5)
net.widget()

# Visualize Specific Cell Type Data

inst_cell_type = 'Sensory neurons'

net.load_df(df_ds)
net.filter_cat(axis='col', cat_index=1, cat_name='Cell Type: ' + inst_cell_type)
net.filter_N_top(inst_rc='row', N_top=250, rank_type='var')
net.normalize(axis='row', norm_type='zscore')
net.clip(-5,5)
net.widget()
